move volume assignments to dedicated file
